+++ system/lib/ex_esdb/cluster.ex
@@ -3,7 +3,6 @@ defmodule ExESDB.Cluster do
   use GenServer
 
   require Logger
-  require BeamCampus.Colors
 
   alias ExESDB.Themes, as: Themes