+++ system/test/beam_campus/colors_test.exs
@@ -1,9 +1,0 @@
-defmodule BeamCampus.ColorsTest do
-  @moduledoc """
-    This module contains tests for the BeamCampus.Colors module.
-  """
-  use ExUnit.Case
-  @tag :bc_colors
-  @tag :bc_docs
-  doctest BeamCampus.Colors
-end